+++ mrsettings/component/getter/component_getter.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/mondegor/go-sysmess/mrmsg"
 	"github.com/mondegor/go-webcore/mrcore"
-	"github.com/mondegor/go-webcore/mrtype"
 
 	"github.com/mondegor/go-components/mrsettings"
 	"github.com/mondegor/go-components/mrsettings/entity"
@@ -32,7 +31,7 @@ func New(parser mrsettings.ValueParser, storage mrsettings.Storage, errorWrapper
 }
 
 // Get - comment method.
-func (co *Component) Get(ctx context.Context, id mrtype.KeyInt32) (string, error) {
+func (co *Component) Get(ctx context.Context, id uint64) (string, error) {
 	value, err := co.getValue(ctx, id, enum.SettingTypeString)
 	if err != nil {
 		return "", err
@@ -47,7 +46,7 @@ func (co *Component) Get(ctx context.Context, id mrtype.KeyInt32) (string, error
 }
 
 // GetList - comment method.
-func (co *Component) GetList(ctx context.Context, id mrtype.KeyInt32) ([]string, error) {
+func (co *Component) GetList(ctx context.Context, id uint64) ([]string, error) {
 	value, err := co.getValue(ctx, id, enum.SettingTypeStringList)
 	if err != nil {
 		return nil, err
@@ -62,7 +61,7 @@ func (co *Component) GetList(ctx context.Context, id mrtype.KeyInt32) ([]string,
 }
 
 // GetInt64 - comment method.
-func (co *Component) GetInt64(ctx context.Context, id mrtype.KeyInt32) (int64, error) {
+func (co *Component) GetInt64(ctx context.Context, id uint64) (int64, error) {
 	value, err := co.getValue(ctx, id, enum.SettingTypeInteger)
 	if err != nil {
 		return 0, err
@@ -77,7 +76,7 @@ func (co *Component) GetInt64(ctx context.Context, id mrtype.KeyInt32) (int64, e
 }
 
 // GetInt64List - comment method.
-func (co *Component) GetInt64List(ctx context.Context, id mrtype.KeyInt32) ([]int64, error) {
+func (co *Component) GetInt64List(ctx context.Context, id uint64) ([]int64, error) {
 	value, err := co.getValue(ctx, id, enum.SettingTypeIntegerList)
 	if err != nil {
 		return nil, err
@@ -92,7 +91,7 @@ func (co *Component) GetInt64List(ctx context.Context, id mrtype.KeyInt32) ([]in
 }
 
 // GetBool - comment method.
-func (co *Component) GetBool(ctx context.Context, id mrtype.KeyInt32) (bool, error) {
+func (co *Component) GetBool(ctx context.Context, id uint64) (bool, error) {
 	value, err := co.getValue(ctx, id, enum.SettingTypeBoolean)
 	if err != nil {
 		return false, err
@@ -106,7 +105,7 @@ func (co *Component) GetBool(ctx context.Context, id mrtype.KeyInt32) (bool, err
 	return parsedValue, nil
 }
 
-func (co *Component) getValue(ctx context.Context, id mrtype.KeyInt32, rowType enum.SettingType) (string, error) {
+func (co *Component) getValue(ctx context.Context, id uint64, rowType enum.SettingType) (string, error) {
 	row, err := co.storage.FetchOne(ctx, id)
 	if err != nil {
 		return "", co.errorWrapper.WrapErrorEntityNotFoundOrFailed(err, entity.ModelNameSetting, mrmsg.Data{"id": id})