+++ mrsort/mrsort.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/mondegor/go-storage/mrentity"
 	"github.com/mondegor/go-storage/mrstorage"
-	"github.com/mondegor/go-webcore/mrtype"
 
 	"github.com/mondegor/go-components/mrsort/entity"
 )
@@ -14,23 +13,23 @@ type (
 	// Orderer - интерфейс управления порядком следования записей.
 	Orderer interface {
 		WithMetaData(meta mrstorage.MetaGetter) Orderer
-		InsertToFirst(ctx context.Context, nodeID mrtype.KeyInt32) error
-		InsertToLast(ctx context.Context, nodeID mrtype.KeyInt32) error
-		MoveToLast(ctx context.Context, nodeID mrtype.KeyInt32) error
-		MoveToFirst(ctx context.Context, nodeID mrtype.KeyInt32) error
-		MoveAfterID(ctx context.Context, nodeID, afterNodeID mrtype.KeyInt32) error
-		Unlink(ctx context.Context, nodeID mrtype.KeyInt32) error
+		InsertToFirst(ctx context.Context, nodeID uint64) error
+		InsertToLast(ctx context.Context, nodeID uint64) error
+		MoveToLast(ctx context.Context, nodeID uint64) error
+		MoveToFirst(ctx context.Context, nodeID uint64) error
+		MoveAfterID(ctx context.Context, nodeID, afterNodeID uint64) error
+		Unlink(ctx context.Context, nodeID uint64) error
 	}
 
 	// Storage - интерфейс для доступа к записям порядка следования и их модификации.
 	Storage interface {
 		WithMetaData(meta mrstorage.MetaGetter) Storage
-		FetchNode(ctx context.Context, nodeID mrtype.KeyInt32) (entity.Node, error)
+		FetchNode(ctx context.Context, nodeID uint64) (entity.Node, error)
 		FetchFirstNode(ctx context.Context) (entity.Node, error)
 		FetchLastNode(ctx context.Context) (entity.Node, error)
 		UpdateNode(ctx context.Context, row entity.Node) error
-		UpdateNodePrevID(ctx context.Context, rowID mrtype.KeyInt32, prevID mrentity.ZeronullInt32) error
-		UpdateNodeNextID(ctx context.Context, rowID mrtype.KeyInt32, nextID mrentity.ZeronullInt32) error
-		RecalcOrderIndex(ctx context.Context, minBorder, step int64) error
+		UpdateNodePrevID(ctx context.Context, rowID uint64, prevID mrentity.ZeronullUint64) error
+		UpdateNodeNextID(ctx context.Context, rowID uint64, nextID mrentity.ZeronullUint64) error
+		RecalcOrderIndex(ctx context.Context, minBorder, step uint64) error
 	}
 )